requirements.txt
Flask
pymongo
cryptography
phe
python-dotenv